# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "EnrollRequest": {
        "description": "A registration.",
        "properties": {
          "student_id": {
            "examples": [
              "S001"
            ],
            "minLength": 1,
            "title": "Student Id",
            "type": "string"
          }
        },
        "required": [
          "student_id"
        ],
        "title": "EnrollRequest",
        "type": "object"
      },
...
      "EnrollmentStatusRequest": {
        "description": "A change of enrolment status.",
        "properties": {
          "status": {
            "description": "`active`, `withdrawn` or `completed`. Withdrawal is a status change rather than a deletion, so grades earned beforehand stay attached.",
            "examples": [
              "withdrawn"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "EnrollmentStatusRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageResponse_CourseResponse_": {
        "properties": {
          "items": {
            "description": "The rows on this page.",
            "items": {
              "$ref": "#/components/schemas/CourseResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "The 1-based page number returned.",
            "title": "Page",
            "type": "integer"
          },
          "pages": {
            "description": "Total number of pages.",
            "title": "Pages",
            "type": "integer"
          },
          "size": {
            "description": "Rows per page.",
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "description": "Rows matching the query overall, ignoring paging.",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "PageResponse[CourseResponse]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nStudent grade management: students, courses, enrolments, grades, reports and an\nAI assistant.\n\n### Errors\n\nEvery failure returns `application/problem+json` (RFC 9457) carrying a stable\nmachine-readable `code` and a structured `context`:\n\n```json\n{ \"type\": \"about:blank#STUDENT_NOT_FOUND\", \"status\": 404,\n  \"code\": \"STUDENT_NOT_FOUND\", \"detail\": \"No student with id 'S999'.\",\n  \"context\": { \"student_id\": \"S999\" } }\n```\n\n**Clients should render the `code`, not the `detail`.** `detail` is a\ndeveloper-facing string and is only ever English; the interface translates the code\ninto the reader's language. This is why the API ships no message catalogue.\n\n### Authorization\n\nAuthentication is a session cookie (`HttpOnly`, `SameSite=Lax`). Access is filtered\nin the query rather than checked in the handler, so what you can see depends on your\nrole:\n\n| Role | Students | Courses | Grades |\n|---|---|---|---|\n| `student` | only themselves | those they are enrolled on | only their own |\n| `teacher` | those enrolled on a course they own | those they own | within their own courses |\n| `admin` | all | all | all |\n| `superadmin` | all, plus AI provider configuration | all | all |\n\nA list endpoint returns what you may see, never a `403`. Only *actions* beyond your\nrole return `403`.\n",
    "title": "Grade Tracker API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/courses": {
      "get": {
        "description": "Paginated, sortable and searchable.\n\nResults are filtered to what your role permits: a student sees only themselves, a teacher sees students enrolled on a course they own, an administrator sees everyone. An out-of-scope id returns `404`, not `403` — a `403` would confirm that a record with that id exists.",
        "operationId": "list_courses_courses_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "description": "`id`, `name`, `term`, `credits` or `created`.",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`id`, `name`, `term`, `credits` or `created`.",
              "title": "Sort"
            }
          },
          {
            "description": "Free text over name and id.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free text over name and id.",
              "title": "Q"
            }
          },
          {
            "description": "Only this academic term.",
            "in": "query",
            "name": "term",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only this academic term.",
              "title": "Term"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_CourseResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List courses",
        "tags": [
          "Courses"
        ]
      },
      "post": {
        "description": "A teacher creating a course owns it by default — the alternative is a course they immediately cannot see.",
        "operationId": "create_course_courses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "`DUPLICATE_ENTRY` — the id is taken."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add a course",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{course_id}": {
      "delete": {
        "description": "Delete a course and, by cascade, its grades and enrolments.",
        "operationId": "delete_course_courses__course_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "description": "`FORBIDDEN` — you do not own this course."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a course",
        "tags": [
          "Courses"
        ]
      },
      "get": {
        "description": "Fetch one course.",
        "operationId": "get_course_courses__course_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "`COURSE_NOT_FOUND` — unknown, or outside your scope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch one course",
        "tags": [
          "Courses"
        ]
      },
      "patch": {
        "description": "Read access is broader than write access: a student can see a course they are enrolled on and cannot rename it. Reassigning `teacher_id` is administrators only, since it changes who can see the course.",
        "operationId": "update_course_courses__course_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "`FORBIDDEN` — you do not own this course."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update a course",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{course_id}/enrollments": {
      "get": {
        "description": "Includes students with no grades yet — `grade_count` of zero. That state is the reason enrolments exist as their own table rather than being inferred from grades.",
        "operationId": "list_enrollments_courses__course_id__enrollments_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentResponse"
                  },
                  "title": "Response List Enrollments Courses  Course Id  Enrollments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "`COURSE_NOT_FOUND`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List a course's register",
        "tags": [
          "Courses"
        ]
      },
      "post": {
        "description": "Enrol a student on a course.",
        "operationId": "enroll_courses__course_id__enrollments_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "`FORBIDDEN` — you do not own this course."
          },
          "409": {
            "description": "`COURSE_FULL` or `DUPLICATE_ENTRY`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enrol a student",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{course_id}/enrollments/{student_id}": {
      "delete": {
        "description": "For correcting a registration made in error. For a student who left, prefer setting the status to `withdrawn` so the record survives.",
        "operationId": "unenroll_courses__course_id__enrollments__student_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "student_id",
            "required": true,
            "schema": {
              "title": "Student Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove an enrolment",
        "tags": [
          "Courses"
        ]
      },
      "patch": {
        "description": "Use `withdrawn` for a student who left; the enrolment row and any grades earned before it are retained.",
        "operationId": "set_enrollment_status_courses__course_id__enrollments__student_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "student_id",
            "required": true,
            "schema": {
              "title": "Student Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollmentStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change an enrolment's status",
        "tags": [
          "Courses"
        ]
      }
    },
...
    "/students": {
      "get": {
        "description": "Paginated, sortable and searchable.\n\nResults are filtered to what your role permits: a student sees only themselves, a teacher sees students enrolled on a course they own, an administrator sees everyone. An out-of-scope id returns `404`, not `403` — a `403` would confirm that a record with that id exists.",
        "operationId": "list_students_students_get",
        "parameters": [
          {
            "description": "1-based page number.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-based page number.",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Rows per page.",
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 25,
              "description": "Rows per page.",
              "maximum": 200,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "description": "`id`, `first_name`, `last_name`, `email` or `created`; `-` to reverse.",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`id`, `first_name`, `last_name`, `email` or `created`; `-` to reverse.",
              "title": "Sort"
            }
          },
          {
            "description": "Free text over name, email and id.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free text over name, email and id.",
              "title": "Q"
            }
          },
          {
            "description": "Only students on this course.",
            "in": "query",
            "name": "course_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only students on this course.",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_StudentResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List students",
        "tags": [
          "Students"
        ]
      },
      "post": {
        "description": "Administrators only. A student record belongs to the institution's register, not to a course — and a teacher creating one could not read it back, since their scope is defined by enrolment. Teachers enrol existing students instead.",
        "operationId": "create_student_students_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "`FORBIDDEN` — administrators only."
          },
          "409": {
            "description": "`DUPLICATE_ENTRY` — the id or email is taken."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add a student",
        "tags": [
          "Students"
        ]
      }
    },
...
    "/students/{student_id}/courses": {
      "get": {
        "description": "List the courses a student is enrolled on.",
        "operationId": "student_courses_students__student_id__courses_get",
        "parameters": [
          {
            "in": "path",
            "name": "student_id",
            "required": true,
            "schema": {
              "title": "Student Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StudentCourseResponse"
                  },
                  "title": "Response Student Courses Students  Student Id  Courses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "`STUDENT_NOT_FOUND`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List a student's courses",
        "tags": [
          "Students"
        ]
      }
    }
  },
...
}